Fix products

+++ TheGioiDiDong/src/main/java/com/tgdd/repository/ProductRepository.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Repository;
 import com.tgdd.entity.Product;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 	@Query("SELECT p FROM Product p where Category_id = :categoryId")
-	 List<Product> getProductbyIdcategory(@Param("categoryId")Integer id);
+	 List<Product> getProductbyIdcategory(@Param("categoryId")Long id);
 	  List<Product> findByProductName(String productName);
-	  Product findByProductId(Integer id);
+	  Product findByProductId(Long id);
 }
